+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Added method `Register<TKey, TService>` to register a service with a specific implementation
   - Added method `Resolve<T>` to resolve a service
   - Added method `Resolve(Type)` to resolve a service
+  - Added method `Clear` to remove any registered service and singleton instance
+  - Added method `Reset` to remove any singleton instance
 - Handle security methods and exceptions
   - When a service is abstract / interface / enum
   - When a service is already registered